requirements.txt
absl-py==0.13.0
astunparse==1.6.3
blis==0.7.4
cached-property==1.5.2
cachetools==4.2.2
catalogue==2.0.4
certifi==2021.5.30
chardet==4.0.0
click==7.1.2
contextvars==2.4
cymem==2.0.5
Cython==0.29.23
dataclasses==0.8
ez-setup==0.9
fastapi==0.66.0
flatbuffers==1.12
gast==0.4.0
google-auth==1.32.1
google-auth-oauthlib==0.4.4
google-pasta==0.2.0
grpcio==1.34.1
h5py==3.1.0
idna==2.10
immutables==0.15
importlib-metadata==4.6.1
Jinja2==3.0.1
joblib==1.0.1
Keras==2.4.3
keras-nightly==2.5.0.dev2021032900
Keras-Preprocessing==1.1.2
Markdown==3.3.4
MarkupSafe==2.0.1
murmurhash==1.0.5
nltk==3.6.2
numpy==1.19.5
oauthlib==3.1.1
opt-einsum==3.3.0
packaging==21.0
pandas==1.1.5
pathy==0.6.0
preshed==3.0.5
protobuf==3.17.3
psycopg2==2.7.7
pyasn1==0.4.8
pyasn1-modules==0.2.8
pydantic==1.8.2
pyparsing==2.4.7
python-dateutil==2.8.1
pytz==2021.1
PyYAML==5.4.1
regex==2021.7.6
requests==2.25.1
requests-oauthlib==1.3.0
rsa==4.7.2
scikit-learn==0.24.2
scipy==1.5.4
six==1.15.0
smart-open==5.1.0
spacy==3.1.0
https://github.com/explosion/spacy-models/releases/download/en_core_web_sm-3.1.0/en_core_web_sm-3.1.0.tar.gz#egg=en_core_web_sm
spacy-legacy==3.0.8
srsly==2.4.1
starlette==0.14.2
tensorboard==2.5.0
tensorboard-data-server==0.6.1
tensorboard-plugin-wit==1.8.0
tensorflow-cpu==2.5.0
tensorflow-estimator==2.5.0
termcolor==1.1.0
thinc==8.0.7
threadpoolctl==2.1.0
tqdm==4.61.2
typer==0.3.2
typing==3.7.4.3
typing-extensions==3.7.4.3
urllib3==1.26.6
wasabi==0.8.2
Werkzeug==2.0.1
wrapt==1.12.1
zipp==3.5.0
py_expression_eval
uvicorn[standard]
pyjwt[crypto]
lark
pyspellchecker
